desktop-client/requirements.txt
customtkinter
websockets
pyautogui
pystray
pillow